Updated the app with Recipe Support. Things to be added: Update Diary UI a bit, Firebase support to backup data online

+++ app/src/main/java/com/example/macrostracker/data/apiModel/FoodApiModel.kt
@@ -21,8 +21,8 @@ fun FoodApiModel.asExternalModel() = Food(
     name = product.productName,
     brand = product.brands,
     calories = product.nutriments.energyKcal100g,
-    carbs = product.nutriments.carbohydrates100g.toInt(),
-    fat = product.nutriments.fat100g.toInt(),
-    protein = product.nutriments.proteins100g.toInt(),
+    carbs = product.nutriments.carbohydrates100g,
+    fat = product.nutriments.fat100g,
+    protein = product.nutriments.proteins100g,
     servingSize = 100
 )
